+++ src/GRA.Domain.Model/Branch.cs
@@ -18,5 +18,7 @@ namespace GRA.Domain.Model
         public string Address { get; set; }
         [MaxLength(50)]
         public string Telephone { get; set; }
+
+        public string SystemName { get; set; }
     }
 }