+++ CHANGELOG.md
@@ -4,8 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
+
 ### Added
 - Add error message when trying to apply a role twice
+
+### Changed
+- Change join branch list to be populated with all branches
 
 ### Fixed
 - Add field to site object to allow forcing https even if the Web server believes the request came in via http.